Numbers have life

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -38,7 +38,7 @@ return (power);
 void print_number(int number)
 {
 int size, digit;
-long counter, sign;
+int counter, sign;
 sign = 1;
 digit = 0;
 size = 1;
